Removes the now obsolete rotator component
