--- src/styles/Rotator.scss
@@ -1,2 +1,0 @@
-.rotator-component {
-}
